Change logic to display ErrorHandler from onDeposit function .
